requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
flask>=2.3.0
lxml>=4.9.0
tqdm>=4.64.0
anthropic>=0.40.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pytesseract>=0.3.10
gunicorn>=21.0.0
